Add: fake and real pull now are helpers
